requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiosqlite==0.19.0
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2
python-multipart==0.0.6
psutil==5.9.6
PyYAML==6.0.1
watchdog==3.0.0
openai